Ignore DotfuscatorReports directory
This contains reports produced when Dotfuscator runs as part of the
integration into MSBuild. Since this is a build output, it shouldn't be
tracked by source control.

Note that these reports can reverse the renaming obfuscation.
As a result, DO NOT DISTRIBUTE THESE REPORTS to end users.
However, you will want to internally archive these reports for each
release, so that if a customer issue contains a stack trace, you can
de-obfuscate the obfuscated stack trace for internal use in diagnosing
the
issue.

# Conflicts:
#	.gitignore